refactor(client-lib): rename `era_fetcher` module to `era`
